[msm.ui]: moved dtraj_states from msm.estimation.obj to msm.ui
