+++ pyemma/msm/ui/dtraj_stats.py
